consolidate log component files
