+++ assets/src/components/cd/logs/LogContent.tsx
