+++ assets/src/components/cd/logs/LogInfo.tsx
