+++ assets/src/components/cd/logs/LogsInfoPanel.tsx
