+++ assets/src/components/cd/logs/LogsLegend.tsx
