+++ assets/src/components/cd/logs/LogsScrollIndicator.tsx
